[Chore] Rename `Processor` to `InputProcessor`
