+++ vllm/model_executor/models/nemotron_vl.py
@@ -34,7 +34,7 @@ from vllm.multimodal import MULTIMODAL_REGISTRY
 from vllm.multimodal.image import convert_image_mode
 from vllm.multimodal.processing import PromptUpdateDetails
 from vllm.sequence import IntermediateTensors
-from vllm.transformers_utils.processor import cached_image_processor_from_config
+from vllm.transformers_utils import cached_image_processor_from_config
 from vllm.transformers_utils.tokenizer import AnyTokenizer
 
 from .interfaces import (
